Update target framework to net7.0

+++ Abc.NCrafts.Quizz/Performance2018/Questions/007/Answer1.cs
@@ -9,7 +9,7 @@ namespace Abc.NCrafts.Quizz.Performance2018.Questions._007
 
         public int Run()
         {
-            var value = _value;
+            var value = (uint)_value;
 
             // begin
             if (Popcnt.IsSupported) // assume is true
